+++ esbuild.js
@@ -4,7 +4,7 @@ const production = process.argv.includes('--production');
 const watch = process.argv.includes('--watch');
 
 async function main() {
-  const ctx = await esbuild.context({
+  const extensionCtx = await esbuild.context({
     entryPoints: ['src/extension.ts'],
     bundle: true,
     format: 'cjs',
@@ -18,12 +18,31 @@ async function main() {
     target: 'es2022',
   });
 
+  const cliCtx = await esbuild.context({
+    entryPoints: ['src/cli.ts'],
+    bundle: true,
+    format: 'cjs',
+    minify: production,
+    sourcemap: !production,
+    sourcesContent: false,
+    platform: 'node',
+    outfile: 'dist/cli.js',
+    logLevel: 'info',
+    target: 'es2022',
+    banner: {
+      js: '#!/usr/bin/env node',
+    },
+  });
+
   if (watch) {
-    await ctx.watch();
+    await extensionCtx.watch();
+    await cliCtx.watch();
     console.log('Watching for changes...');
   } else {
-    await ctx.rebuild();
-    await ctx.dispose();
+    await extensionCtx.rebuild();
+    await cliCtx.rebuild();
+    await extensionCtx.dispose();
+    await cliCtx.dispose();
   }
 }